[8/27][41/183] Add answers about vectors!
Treating bit arrays in CL is so nice! I liked to do the
list-to-bit-vector function! Really nice. We declare #*0101
literal arrays that way, in that case is a bit-vector (a type
defined in common lisp)!

But we can try others like #(1 2 3) and the lisp machine
will infer the type!

We have some primitive to do bitwise with bit-vectorss like:
(bit-and #*1000 #*1111) ->  #1000
(bit-ior #*1000 #*1111) ->  #1111
(bit-xor #*1000 #*1111) ->  #0111
(bit-nand #*1000 #*1111) -> #0111

Check in:
https://www.cs.cmu.edu/Groups/AI/html/cltl/clm/node161.html

Remember on make-array use the keyword argument :element-type
declaring explicitly 'bit for that kind of stuff with
bit-vectors.