[28/28][188/188] Add answers of koans about threads! FINISHED ALL!!!
Well, after about one year I finished this stuff. The last topic is threads
using the SBCL compiler. A special note on this koans is written on the header:
the threads management is made as compiler/interpreter specific, then
this is not covered into the Common Lisp general standards.

Seems SBCL has great features to control threads, like mutex and semaphores.

A collection of useful statements for handling threads are:
(namespace of threads => :sb-thread)

* sb-thread:make-thread
  * the first argument is a function
  * use :arguments to pass a list of arguments to apply a function
  * use :name to give a name to the thread (useful for debug)
* sb-thread:join-thread
  * sync the thread
* sb-thread:terminate-thread
  * kill a given thread as argument
* sb-thread:list-all-threads
  * return a list of the current active threads
* sb-thread:make-mutex
  * create a mutex, can be used with the lexical lock
      (sb-thread:with-mutex *mutex*
                             ... body)
  * use :name to given the mutex a name
* sb-thread:make-semaphore
  * create a semaphore
  * use :name to given it a name
  * use :count to given it a initial count
  * use (sb-thread:wait-on-semaphore *s*) until :count > 0
  * use (sb-thread:semaphore-count *s*) to increase the :count slot

Mutex => thread lockers
Semaphores => counters with atomic operations

NOTE:
Semaphores in general is useful on the pattern problem of consumer and producer;
on the last test of threads.lsp is defined a problem like that. As well, I
give the answers. Is used a mutex to control access of a shared vector and
one semaphore to alert the threads about production of apples and permission
to eaten them.

I need comment the problematic test of shared global variable accumulator
with threads. This give to us a race condition and a undefined behavior. A
equivalent example using mutex is given after this. (to keep the global count
of tests unmodified I created a new stupid test for this)

After complete this series of koans, the script give me that:

**********************************************************
That was the last one, well done!  ENLIGHTENMENT IS YOURS!
**********************************************************